Depend on nom with `default-features = false` (#47)
* Update nom dependency in Cargo.toml

This change is necessary to allow nom_locate to be used without using the default features of nom, e.g. "lexical".

* Change nom dependency to require "std" feature.

Since a lot of stuff in nom_locate depends on nom having the "std" feature, it probably makes sense to require this feature. Other nom features such as "regexp" and "lexical" will still be optional.

* Add #[cfg(feature = alloc)] to tests which require the nom alloc feature.

* Revert "Add #[cfg(feature = alloc)] to tests which require the nom alloc feature."

This reverts commit 6b72f8b99783a9b40aaa93a3dbc421a2f5f08fbd.

* Remove "std" feature from nom dependency, and allow failures with "alloc" on stable/beta.

* Add removed jobs back to travis.yml include section.

* Add missing quote

* Change #[cfg(feature = "alloc")] to #[cfg(any(feature = "std", feature = "alloc"))]